# backend-go/internal/infrastructure/http/apidocs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Course AI API",
    "version": "1.0.0",
    "description": "API HTTP du MVP Course AI protégée par des sessions Clerk. La génération complète est asynchrone et s'appuie sur des jobs PostgreSQL durables."
  },
  "servers": [
    {
      "url": "/",
      "description": "Hôte courant"
    }
  ],
...
  "paths": {
...
    "/api/courses": {
      "get": {
        "tags": [
          "Courses"
        ],
        "operationId": "listCourses",
        "summary": "Lister les formations",
        "description": "Retourne une page de formations triées. pageSize est normalisé par le backend et limité à 100.",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "Recherche textuelle insensible à la casse dans le titre ou le synopsis.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filtre sur l'état de génération du cours.",
            "schema": {
              "$ref": "#/components/schemas/CourseStatus"
            }
          },
          {
            "name": "language",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Language"
            }
          },
          {
            "name": "orderBy",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "created_at",
                "updated_at",
                "title",
                "status"
              ],
              "default": "created_at"
            }
          },
          {
            "name": "orderDirection",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "desc"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Page de formations.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoursePageResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/courses/{courseID}": {
      "get": {
        "tags": [
          "Courses"
        ],
        "operationId": "getCourse",
        "summary": "Récupérer une formation complète",
        "description": "Charge le cours et toutes ses relations : modules, leçons, exercices et quiz.",
        "parameters": [
          {
            "$ref": "#/components/parameters/CourseID"
          }
        ],
        "responses": {
          "200": {
            "description": "Formation trouvée.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "delete": {
        "tags": [
          "Courses"
        ],
        "operationId": "deleteCourse",
        "summary": "Supprimer une formation",
        "description": "Supprime le cours et ses relations selon les contraintes de cascade PostgreSQL.",
        "parameters": [
          {
            "$ref": "#/components/parameters/CourseID"
          }
        ],
        "responses": {
          "204": {
            "description": "Formation supprimée, sans corps de réponse."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/courses/{courseID}/modules": {
      "get": {
        "tags": [
          "Courses",
          "Modules"
        ],
        "operationId": "listCourseModules",
        "summary": "Lister les modules d'une formation",
        "parameters": [
          {
            "$ref": "#/components/parameters/CourseID"
          }
        ],
        "responses": {
          "200": {
            "description": "Modules ordonnés du cours.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ModuleResponse"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/modules/{moduleID}": {
      "get": {
        "tags": [
          "Modules"
        ],
        "operationId": "getModule",
        "summary": "Récupérer un module",
        "parameters": [
          {
            "$ref": "#/components/parameters/ModuleID"
          }
        ],
        "responses": {
          "200": {
            "description": "Module trouvé avec ses leçons.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/modules/{moduleID}/lessons": {
      "get": {
        "tags": [
          "Modules",
          "Lessons"
        ],
        "operationId": "listModuleLessons",
        "summary": "Lister les leçons d'un module",
        "parameters": [
          {
            "$ref": "#/components/parameters/ModuleID"
          }
        ],
        "responses": {
          "200": {
            "description": "Leçons ordonnées du module.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LessonResponse"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/lessons/{lessonID}": {
      "get": {
        "tags": [
          "Lessons"
        ],
        "operationId": "getLesson",
        "summary": "Récupérer une leçon complète",
        "description": "Retourne le Markdown, les exercices structurés et les quiz de la leçon.",
        "parameters": [
          {
            "$ref": "#/components/parameters/LessonID"
          }
        ],
        "responses": {
          "200": {
            "description": "Leçon trouvée.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/lessons/{lessonID}/solutions": {
      "get": {
        "tags": ["Lessons"],
        "operationId": "getLessonSolutions",
        "summary": "Révéler les solutions d'une leçon",
        "description": "Retourne explicitement les corrections des exercices et les réponses des quiz. Les endpoints ordinaires n'exposent jamais ces données.",
        "parameters": [{"$ref": "#/components/parameters/LessonID"}],
        "responses": {
          "200": {
            "description": "Solutions de la leçon.",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/LessonSolutionsResponse"}}}
          },
          "400": {"$ref": "#/components/responses/BadRequest"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "500": {"$ref": "#/components/responses/InternalError"}
        }
      }
    }
  },
  "components": {
    "parameters": {
      "CourseID": {
        "name": "courseID",
        "in": "path",
        "required": true,
        "description": "Identifiant UUID du cours.",
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "ModuleID": {
        "name": "moduleID",
        "in": "path",
        "required": true,
        "description": "Identifiant UUID du module.",
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "LessonID": {
        "name": "lessonID",
        "in": "path",
        "required": true,
        "description": "Identifiant UUID de la leçon.",
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
    },
    "responses": {
...
      "BadRequest": {
        "description": "Corps, paramètre ou règle de validation invalide.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "code": "validation_error",
              "message": "invalid level"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Session Clerk absente ou invalide.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "code": "unauthenticated",
              "message": "authentication is required"
            }
          }
        }
      },
      "NotFound": {
        "description": "Ressource absente.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "code": "course_not_found",
              "message": "course not found"
            }
          }
        }
      },
...
      "ServiceUnavailable": {
        "description": "Une dépendance applicative nécessaire est indisponible.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "code": "service_unavailable",
              "message": "generation service is unavailable"
            }
          }
        }
      },
...
      "InternalError": {
        "description": "Erreur interne masquée au client et journalisée avec sa stack trace.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "code": "internal_error",
              "message": "internal server error"
            }
          }
        }
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Code stable exploitable par le client."
          },
          "message": {
            "type": "string",
            "description": "Message public sans détail interne."
          }
        }
      },
...
      "CoursePageResponse": {
        "type": "object",
        "required": [
          "items",
          "page",
          "pageSize",
          "totalItems",
          "totalPages",
          "hasNext",
          "hasPrevious"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CourseSummaryResponse"
            }
          },
          "page": {
            "type": "integer",
            "minimum": 1
          },
          "pageSize": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          },
          "totalItems": {
            "type": "integer",
            "minimum": 0
          },
          "totalPages": {
            "type": "integer",
            "minimum": 0
          },
          "hasNext": {
            "type": "boolean"
          },
          "hasPrevious": {
            "type": "boolean"
          }
        }
      },
      "CourseSummaryResponse": {
        "type": "object",
        "required": ["id", "requestId", "language", "status", "title", "synopsis", "currentLevel", "targetLevel", "createdAt", "updatedAt"],
        "properties": {
          "id": {"type": "string", "format": "uuid"},
          "requestId": {"type": "string", "format": "uuid"},
          "language": {"$ref": "#/components/schemas/Language"},
          "status": {"$ref": "#/components/schemas/CourseStatus"},
          "title": {"type": "string"},
          "synopsis": {"type": "string"},
          "currentLevel": {"$ref": "#/components/schemas/Level"},
          "targetLevel": {"$ref": "#/components/schemas/Level"},
          "createdAt": {"type": "string", "format": "date-time"},
          "updatedAt": {"type": "string", "format": "date-time"}
        }
      },
      "CourseResponse": {
        "type": "object",
        "required": [
          "id",
          "requestId",
          "language",
          "status",
          "initialUserPrompt",
          "title",
          "synopsis",
          "targetAudience",
          "currentLevel",
          "targetLevel",
          "prerequisites",
          "goals",
          "acquiredSkills",
          "finalProjectTitle",
          "finalProjectDescription",
          "finalProjectConstraints",
          "totalDurationMinutes",
          "modules",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "language": {
            "$ref": "#/components/schemas/Language"
          },
          "status": {
            "$ref": "#/components/schemas/CourseStatus"
          },
          "initialUserPrompt": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "synopsis": {
            "type": "string"
          },
          "targetAudience": {
            "type": "string",
            "nullable": true
          },
          "currentLevel": {
            "$ref": "#/components/schemas/Level"
          },
          "targetLevel": {
            "$ref": "#/components/schemas/Level"
          },
          "prerequisites": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "goals": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "acquiredSkills": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "finalProjectTitle": {
            "type": "string",
            "nullable": true
          },
          "finalProjectDescription": {
            "type": "string",
            "nullable": true
          },
          "finalProjectConstraints": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "totalDurationMinutes": {
            "type": "integer",
            "minimum": 0,
            "description": "Somme calculée des durées estimées des leçons."
          },
          "modules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModuleResponse"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ModuleResponse": {
        "type": "object",
        "required": [
          "id",
          "courseId",
          "order",
          "title",
          "description",
          "keyLearningPoints",
          "totalDurationMinutes",
          "lessons",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "courseId": {
            "type": "string",
            "format": "uuid"
          },
          "order": {
            "type": "integer",
            "minimum": 1
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "keyLearningPoints": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "totalDurationMinutes": {
            "type": "integer",
            "minimum": 0,
            "description": "Somme calculée des durées estimées des leçons du module."
          },
          "lessons": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LessonResponse"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "LessonResponse": {
        "type": "object",
        "required": [
          "id",
          "moduleId",
          "order",
          "title",
          "type",
          "estimatedDurationMinutes",
          "learningGoal",
          "requiresDiagram",
          "technicalKeywords",
          "contentMarkdown",
          "exercises",
          "quizzes",
          "hasContent",
          "hasStructuredActivities",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "moduleId": {
            "type": "string",
            "format": "uuid"
          },
          "order": {
            "type": "integer",
            "minimum": 1
          },
          "title": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/LessonType"
          },
          "estimatedDurationMinutes": {
            "type": "integer",
            "minimum": 1
          },
          "learningGoal": {
            "type": "string"
          },
          "requiresDiagram": {
            "type": "boolean"
          },
          "technicalKeywords": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "contentMarkdown": {
            "type": "string",
            "nullable": true
          },
          "exercises": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExerciseResponse"
            }
          },
          "quizzes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuizResponse"
            }
          },
          "hasContent": {
            "type": "boolean"
          },
          "hasStructuredActivities": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ExerciseResponse": {
        "type": "object",
        "required": [
          "id",
          "lessonId",
          "type",
          "difficulty",
          "title",
          "objective",
          "instructionsMarkdown",
          "contentMarkdown",
          "payload",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "lessonId": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "$ref": "#/components/schemas/ExerciseType"
          },
          "difficulty": {
            "$ref": "#/components/schemas/Difficulty"
          },
          "title": {
            "type": "string"
          },
          "objective": {
            "type": "string"
          },
          "instructionsMarkdown": {
            "type": "string"
          },
          "contentMarkdown": {
            "type": "string"
          },
          "payload": {
            "$ref": "#/components/schemas/ExercisePayloadResponse"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ExercisePayloadResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["tasks", "resources", "starterCode", "expectedOutput", "hints"],
        "properties": {
          "tasks": {"type": "array", "items": {"type": "string"}},
          "resources": {"type": "array", "items": {"type": "string"}},
          "starterCode": {"type": "string", "nullable": true},
          "expectedOutput": {"type": "string", "nullable": true},
          "hints": {"type": "array", "items": {"type": "string"}}
        }
      },
      "QuizResponse": {
        "type": "object",
        "required": [
          "id",
          "lessonId",
          "type",
          "difficulty",
          "title",
          "objective",
          "questions",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "lessonId": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "$ref": "#/components/schemas/QuizType"
          },
          "difficulty": {
            "$ref": "#/components/schemas/Difficulty"
          },
          "title": {
            "type": "string"
          },
          "objective": {
            "type": "string"
          },
          "questions": {
            "type": "array",
            "minItems": 1,
            "items": {
              "$ref": "#/components/schemas/QuizQuestionResponse"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "QuizQuestionResponse": {
        "type": "object",
        "required": [
          "order",
          "type",
          "question",
          "options"
        ],
        "properties": {
          "order": {
            "type": "integer",
            "minimum": 1
          },
          "type": {
            "$ref": "#/components/schemas/QuizQuestionType"
          },
          "question": {
            "type": "string"
          },
          "options": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuizOptionResponse"
            }
          }
        }
      },
      "QuizOptionResponse": {
        "type": "object",
        "required": [
          "order",
          "text"
        ],
        "properties": {
          "order": {
            "type": "integer",
            "minimum": 1
          },
          "text": {
            "type": "string"
          }
        }
      },
      "QuizAnswerResponse": {
        "type": "object",
        "required": [
          "answer",
          "answers"
        ],
        "properties": {
          "answer": {
            "type": "string",
            "nullable": true
          },
          "answers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "LessonSolutionsResponse": {
        "type": "object",
        "required": ["lessonId", "exercises", "quizzes"],
        "properties": {
          "lessonId": {"type": "string", "format": "uuid"},
          "exercises": {
            "type": "array",
            "items": {"$ref": "#/components/schemas/ExerciseSolutionResponse"}
          },
          "quizzes": {
            "type": "array",
            "items": {"$ref": "#/components/schemas/QuizSolutionResponse"}
          }
        }
      },
      "ExerciseSolutionResponse": {
        "type": "object",
        "required": ["exerciseId", "correctionMarkdown"],
        "properties": {
          "exerciseId": {"type": "string", "format": "uuid"},
          "correctionMarkdown": {"type": "string"}
        }
      },
      "QuizSolutionResponse": {
        "type": "object",
        "required": ["quizId", "questions"],
        "properties": {
          "quizId": {"type": "string", "format": "uuid"},
          "questions": {
            "type": "array",
            "items": {"$ref": "#/components/schemas/QuizQuestionSolutionResponse"}
          }
        }
      },
      "QuizQuestionSolutionResponse": {
        "type": "object",
        "required": ["order", "answer", "correction"],
        "properties": {
          "order": {"type": "integer", "minimum": 1},
          "answer": {"$ref": "#/components/schemas/QuizAnswerResponse"},
          "correction": {"type": "string"}
        }
      },
      "Language": {
        "type": "string",
        "enum": [
          "fr",
          "en"
        ]
      },
      "Level": {
        "type": "string",
        "enum": [
          "beginner",
          "intermediate",
          "advanced",
          "expert",
          "unknown"
        ]
      },
      "LessonType": {
        "type": "string",
        "enum": [
          "theory",
          "practice",
          "mixed",
          "quiz"
        ]
      },
      "Difficulty": {
        "type": "string",
        "enum": [
          "beginner",
          "intermediate",
          "advanced"
        ]
      },
      "ExerciseType": {
        "type": "string",
        "enum": [
          "guided_lab",
          "coding",
          "debugging",
          "configuration",
          "scenario",
          "written_answer",
          "command_line",
          "mixed"
        ]
      },
      "QuizType": {
        "type": "string",
        "enum": [
          "single_choice",
          "multiple_choice",
          "true_false",
          "short_answer",
          "mixed"
        ]
      },
      "QuizQuestionType": {
        "type": "string",
        "enum": [
          "single_choice",
          "multiple_choice",
          "true_false",
          "short_answer"
        ]
      },
...
      "CourseStatus": {
        "type": "string",
        "enum": [
          "analysis_pending",
          "analysis_completed",
          "architecture_generating",
          "structure_generated",
          "lessons_generating",
          "lessons_generated",
          "content_generating",
          "completed",
          "failed"
        ]
      },
...
    },
    "securitySchemes": {
      "ClerkBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Clerk session JWT",
        "description": "Session token Clerk transmis dans Authorization: Bearer <token>."
      }
    }
  },
  "security": [
    {
      "ClerkBearer": []
    }
  ]
}